:fire: PRTL-1866 replace react-highlighter with our own
