+++ package.json
@@ -58,7 +58,6 @@
     "react-dom": "16.0.0",
     "react-faux-dom": "3.1.0",
     "react-helmet": "5.1.3",
-    "react-highlighter": "0.4.0",
     "react-icons": "2.2.5",
     "react-loadable": "4.0.2",
     "react-modal": "3.0.0",
